// web/src/schema/site.schema.d.ts
export interface SAMLAuthProvider {
  type: string;
  /**
   * An identifier that can be used to reference this authentication provider in other parts of the config. For example, in configuration for a code host, you may want to designate this authentication provider as the identity provider for the code host.
   */
  configID?: string;
  /**
   * The name to use when displaying this authentication provider in the UI. Defaults to an auto-generated name with the type of authentication provider and other relevant identifiers (such as a hostname).
   */
  displayName?: string;
  /**
   * The name of this SAML Service Provider, which is used by the Identity Provider to identify this Service Provider. It defaults to https://sourcegraph.example.com/.auth/saml/metadata (where https://sourcegraph.example.com is replaced with this Sourcegraph instance's "externalURL"). It is only necessary to explicitly set the issuer if you are using multiple SAML authentication providers.
   */
  serviceProviderIssuer?: string;
  /**
   * The SAML Identity Provider metadata URL (for dynamic configuration of the SAML Service Provider).
   */
  identityProviderMetadataURL?: string;
  /**
   * The SAML Identity Provider metadata XML contents (for static configuration of the SAML Service Provider). The value of this field should be an XML document whose root element is `<EntityDescriptor>` or `<EntityDescriptors>`.
   */
  identityProviderMetadata?: string;
  /**
   * The SAML Service Provider certificate in X.509 encoding (begins with "-----BEGIN CERTIFICATE-----"). This certificate is used by the Identity Provider to validate the Service Provider's AuthnRequests and LogoutRequests. It corresponds to the Service Provider's private key (`serviceProviderPrivateKey`).
   */
  serviceProviderCertificate?: string;
  /**
   * The SAML Service Provider private key in PKCS#8 encoding (begins with "-----BEGIN PRIVATE KEY-----"). This private key is used to sign AuthnRequests and LogoutRequests. It corresponds to the Service Provider's certificate (`serviceProviderCertificate`).
   */
  serviceProviderPrivateKey?: string;
  /**
   * The SAML NameID format to use when performing user authentication.
   */
  nameIDFormat?: string;
  /**
   * Sign AuthnRequests and LogoutRequests sent to the Identity Provider using the Service Provider's private key (`serviceProviderPrivateKey`). It defaults to true if the `serviceProviderPrivateKey` and `serviceProviderCertificate` are set, and false otherwise.
   */
  signRequests?: boolean;
  /**
   * Whether the Service Provider should (insecurely) accept assertions from the Identity Provider without a valid signature.
   */
  insecureSkipAssertionSignatureValidation?: boolean;
}
/**
 * Configures the OpenID Connect authentication provider for SSO.
 */
export interface OpenIDConnectAuthProvider {
  type: string;
  /**
   * The name to use when displaying this authentication provider in the UI. Defaults to an auto-generated name with the type of authentication provider and other relevant identifiers (such as a hostname).
   */
  displayName?: string;
  /**
   * An identifier that can be used to reference this authentication provider in other parts of the config. For example, in configuration for a code host, you may want to designate this authentication provider as the identity provider for the code host.
   */
  configID?: string;
  /**
   * The URL of the OpenID Connect issuer.
   *
   * For Google Apps: https://accounts.google.com
   */
  issuer: string;
  /**
   * The client ID for the OpenID Connect client for this site.
   *
   * For Google Apps: obtain this value from the API console (https://console.developers.google.com), as described at https://developers.google.com/identity/protocols/OpenIDConnect#getcredentials
   */
  clientID: string;
  /**
   * The client secret for the OpenID Connect client for this site.
   *
   * For Google Apps: obtain this value from the API console (https://console.developers.google.com), as described at https://developers.google.com/identity/protocols/OpenIDConnect#getcredentials
   */
  clientSecret: string;
  /**
   * Only allow users to authenticate if their email domain is equal to this value (example: mycompany.com). Do not include a leading "@". If not set, all users on this OpenID Connect provider can authenticate to Sourcegraph.
   */
  requireEmailDomain?: string;
}